src/services/chatterbox/requirements.txt
chatterbox-tts>=0.1.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
torchaudio>=2.6.0
torch>=2.6.0
librosa>=0.10.0
numpy>=1.26.0
